Bundle post-install instead of pre-publish!
The version is native-buffer-browserify is currently stuck at whatever
it was when you last published to npm.

For example, right now ‘insert-module-globals’ is using
native-buffer-browserify@2.0.0 but ‘browserify’ is using 2.0.13.

I know post-install scripts are usually frowned upon, but getting two
different versions of Buffer (one with bug fixes and one without) is
worse.